Add client management features and update dependencies
- Implement client registration functionality
- Refactor menu to handle client list
- Update `Cargo.toml` and `Cargo.lock` with new dependencies
- Introduce utility functions for screen clearing and waiting

+++ src/models/cliente.rs
@@ -1,7 +1,8 @@
 
 #[allow(dead_code)]
+#[derive(Default)]
 pub struct Cliente {
-    pub id: i32,
+    pub id: usize,
     pub nome: String,
     pub cpf: String,
     pub endereco: String,
